requirements.txt
Django>=5.0,<6
PyPDF2>=3.0.0
Pillow>=10.0.0
PyMuPDF>=1.24.0
